src: Add CPPC tests
This requires "acpi_cppc" driver and "schedutils" governor. It will
assume the latest online CPU running at the lowest frequency, and then
force it to run at the highest frequency. Finally, it should scale down
once the workload is gone. We will check performance counters to confirm
them as well.

Signed-off-by: Qian Cai <quic_qiancai@quicinc.com>