src: Add a PCIe test
It will try to read all files under PCIe sysfs. If the directory is a
device path, we will skip it as we don't want to deal with device's
specific file return errors. Otherwise, we will check to see if the file
is accessible. If so, we will attempt to read it and compare with the
allow list for the errors. Finally, we will print out the list of
unexpected failures.

Signed-off-by: Qian Cai <quic_qiancai@quicinc.com>